CA-248332: Fix race condition in iostat while updating VDI-to-VM map.
There is a window when the tapdisk is created but the VBDs are not,
and if VDI-to-VM map is updated during this window then the daemon will not
try to update the map again until the next tapdisk is created.

This fix will call `update_vdi_to_vm_map` if any of the tapdisk
known doesn't have corresponding entries in the VDI-to-VM map.

Signed-off-by: Sharad Yadav <sharad.yadav@citrix.com>