XSI-400 Backport: stop xcp-rrdd-isostat log spam
Updating our mapping between VDIs and VMs appears
to spam the logs the most, so we try to reduce
occurences of this here.

We compare the latest VDIs (coming from the output
of tap-ctl list) to the VDIs in the previous
mapping to see if we actually do need to update
the map. We update the mapping regardless every
five minutes to avoid stale data.

Signed-off-by: lippirk <ben.anson@citrix.com>