CA-325582: Report stats about unplugged SRs
This is needed  because SR metrics are being reported as host ones:
when creating and deleting SRs their metrics data-sources are kept
and accumulate over time.
To avoid an accompanying change in xcp-rrdd deletes rrds with no
corresponding active data-sources for host metrics.

To avoid the deletion of metrics of unplugged PBDS, we make an
extra effort to report them.
This means querying xapi for unplugged PBDs and stuffing metrics with
unknown values to not upset the existing, valid metrics in the SRs.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>